Start separating files: UNFINISHED!
